Add files via upload
hello guys, in my adventure to learn how to code I am trying to connect my program to a website and be able to be able to make the user input a name and then the website takes that input and is able to make a wutang clan name from that input. The file I have uploaded right now, 7/27/2021, is only the beginning.